working operator, clear and equal buttons
The clear button empties the display and all variables used for
operations.

The equal button calls the clear function first and then shows the
calculation result on the display. Note that all variables are cleared.

If variable a has a value, the operator buttons act as an equal button
first, showing the calculation result on the display.
If it doesn't it will the user input is already on the display.
After that, it stores values for the operator and a variables.